Change 'long long' to 'int64_t'.
BUG=none
TEST=none
TBR=darin

Review URL: http://codereview.chromium.org/3944001

+++ c/dev/ppb_file_system_dev.h
@@ -22,7 +22,7 @@ struct PPB_FileSystem_Dev {
   // Opens the file system. A file system must be opened before running any
   // other operation on it.
   int32_t (*Open)(PP_Resource file_system,
-                  long long expected_size,
+                  int64_t expected_size,
                   struct PP_CompletionCallback callback);
 };
